separate FollowRepository into its interface and implementation
